sdm: add property to skip creating extension interface
For some platforms, the extension interface requires
libraries that are not provided, like libscalar.so.

To work around the issue, simply skip loading the
extension library and the interface creation. This
does mean losing the possibility to have SDE
composition, but the platforms that need this never
used it previously, so it is not a critical loss.

Change-Id: I0900b4b5fffb21353bc27141e1356bb08cd2ba4a
Signed-off-by: Chippa-a <bagirovvusal99@gmail.com>